[lib32-sni-qt] Use 32-bit qconfig.h
The plugin produced by the build contains a x86-64 build key as it is
built against the "qconfig.h" file produced by the x86-64 build for the
host system. In turn, the 32-bit build of Qt refuses to load the plugin
as the build key does not match.

Fix this by providing a custom "qconfig.h" file during the build and
prepending its location in the preprocessor's include search path. This
is the same approach chosen in the "lib32-qtcurve" package which is part
of the multilib repository.